+++ src/Enums/ParseType.php
@@ -10,8 +10,6 @@ enum ParseType: string
 
     case JSON = 'json';
 
-    //case OPENAPI = 'openapi';
-
     case CONSOLE = 'console';
 
     case MODEL = 'model';
@@ -19,11 +17,20 @@ enum ParseType: string
     public function toString(): string
     {
         return match ($this) {
-            self::TABLE => 'table',
-            self::JSON => 'json',
-            self::CONSOLE => 'console',
-            self::MODEL => 'model',
-            //self::OPENAPI => 'openapi yaml (beta)',
+            self::TABLE     => 'table',
+            self::JSON      => 'json',
+            self::CONSOLE   => 'console',
+            self::MODEL     => 'model',
+        };
+    }
+
+    public function getCommandName(): string
+    {
+        return match ($this) {
+            self::TABLE     => 'moonshine:build-table',
+            self::JSON      => 'moonshine:build-json',
+            self::CONSOLE   => 'moonshine:build-resource',
+            self::MODEL     => 'moonshine:build-model',
         };
     }
 }
